requirements.txt
confluent-kafka==2.0.2
python-dotenv==0.21.1
